Spring MVC added

+++ src/main/java/io/spring/boot/service/UserServiceJpaRepositoryImpl.java
@@ -28,7 +28,7 @@ public class UserServiceJpaRepositoryImpl implements UserService {
 
     @Override
     public User findById(Long id) {
-        return userRepository.findById(id).orElse(null);
+        return userRepository.findById(id.intValue()).orElse(null);
     }
 
     @Override
